front-end-main/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
PyMySQL
sqlalchemy-utils
pandas
selenium
webdriver-manager
beautifulsoup4